% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evolutionary_parameter_gibbs.R
\name{coal_rate_gibbs_update}
\alias{coal_rate_gibbs_update}
\alias{bit_mig_mat_gibbs_update}
\title{Evolutionary Parameter Gibbs Updates}
\usage{
coal_rate_gibbs_update(strphylo, n_deme, prior_shape = 1, prior_rate = 1)

bit_mig_mat_gibbs_update(strphylo, n_deme, prior_shape = 1, prior_rate = 1)
}
\arguments{
\item{strphylo}{Dated structured phylogeny}

\item{prior_shape}{Gamma shape parameter for prior}

\item{prior_rate}{Gamma rate parameter for prior}
}
\value{
Updated parameters
}
\description{
Gibbs updates for evolutionary parameters governing the structured coalescent
}
