% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ewing_2004_proposals.R
\name{coalescent_merge}
\alias{coalescent_merge}
\title{Coalescent Node Merge Proposal}
\usage{
coalescent_merge(ED, n_deme, node_indices)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n_deme}{Number of distinct demes in the population}

\item{node_indices}{Vector giving row indices for node labels}
}
\value{
Updated extended data object with the proposal from the migration pair birth move
}
\description{
Performs a coalescent node merge move (Ewing et al. 2004). Merges two migration
nodes immediately below a coalescent node and place above coalescent node
}
