% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ewing_2004_proposals.R
\name{coalescent_split}
\alias{coalescent_split}
\title{Coalescent Node Split Proposal}
\usage{
coalescent_split(ED, n_deme, node_indices)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n_deme}{Number of distinct demes in the population}

\item{node_indices}{Vector giving row indices for node labels}
}
\value{
Updated extended data object with the proposal from the migration pair birth move
}
\description{
Performs a coalescent node split move (Ewing et al. 2004). Splits a migration
node immediately above a coalescent node into two migration nodes placed immediately
below the coalescent node
}
