% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Default_priors.R
\name{default_priors}
\alias{default_priors}
\title{Default StructCoalescent prior parameters}
\usage{
default_priors(strphylo, n_deme, M = fitch(strphylo)$min_migs)
}
\arguments{
\item{strphylo}{strphylo object giving an initial structured phylogeny}

\item{n_deme}{Number of demes in the problem}

\item{M}{(optional) prior target number of migration events in a migration history. Default value places prior mass on maximum parsimony migration histories}
}
\value{
Updated extended data object with the proposal from the migration birth move
}
\description{
Computes the prior parameters for the default prior specification for StructCoalescent
}
