% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{dexp_trunc}
\alias{dexp_trunc}
\alias{pexp_trunc}
\alias{rexp_trunc}
\title{The Truncated Exponential Distribution}
\usage{
dexp_trunc(x, c = Inf, rate = 1)

pexp_trunc(q, c = Inf, rate = 1)

rexp_trunc(n, c = Inf, rate = 1)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{c}{truncation cutoff}

\item{rate}{vector of rates}

\item{n}{number of observations}
}
\description{
Density, distribution function and random generation for the exponential
distribution with rate \code{rate} truncated above by \code{c}
}
