% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{DTA_likelihood}
\alias{DTA_likelihood}
\title{DTA likelihood}
\usage{
DTA_likelihood(
  strphylo,
  fit_mig_mat = FitMigMatC(bit_mig_mat, coal_rate),
  bit_mig_mat = BitMigMatC(fit_mig_mat, coal_rate),
  coal_rate
)
}
\arguments{
\item{strphylo}{Structured dated phylogeny}

\item{fit_mig_mat}{Matrix of forwards-in-time migration rates between pairs of demes}

\item{bit_mig_mat}{Matrix of backwards-in-time migration rates between pairs of demes}

\item{coal_rate}{Vector of coalescent rate in each deme (1 / effective population size)}
}
\value{
List containing log-likelihood and likelihood of the structured phylogenetic tree
}
\description{
Computes the likelihood of a structured phylogenetic tree under the Discrete Trait Analysis (DTA) model
}
