% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DTALikelihoodC}
\alias{DTALikelihoodC}
\title{DTALikelihoodC}
\usage{
DTALikelihoodC(ED, fit_mig_mat, node_indices)
}
\arguments{
\item{ED}{NumericMatrix Extended data object representing structured phylogeny}

\item{node_indices}{NumericVector Vector giving row numbers in ED for node labels}

\item{coal_rate}{NumericVector Coalescent rate in each deme (1 / effective population size)}

\item{bit_mig_mat}{NumericMatrix Backwards-in-time migration rates between pairs of demes}
}
\value{
List containing log-likelihood and likelihood of the structured coalescent genealogy
}
\description{
Computes the DTA likelihood of a structured genealogy
}
