% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{Ewing_2004_mcmc}
\alias{Ewing_2004_mcmc}
\title{Ewing et al. (2004) MCMC}
\usage{
Ewing_2004_mcmc(
  N,
  strphylo,
  coal_rate,
  bit_mig_mat,
  cr_mode,
  cr_var,
  mm_mode,
  mm_var,
  output_dir,
  run_name = "Ewing_2004",
  stdout_log = TRUE,
  thin = max(N/5000, 1),
  save_migration_history = TRUE,
  migration_history_thin = thin,
  proposal_rates = c(rep(100, 6), 1, 1)
)
}
\arguments{
\item{N}{Total number of MCMC iterations to complete}

\item{strphylo}{Initial structured phylogenetic tree}

\item{coal_rate}{Initial estimate of coalescent rates}

\item{bit_mig_mat}{Initial estimate of backward-in-time migration rates matrix}

\item{cr_mode}{Coalescent rates prior mode}

\item{cr_var}{Coalescent rates prior variance}

\item{mm_mode}{Backward-in-time migration rates prior mode}

\item{mm_var}{Backward-in-time migration rates prior variance}

\item{output_dir}{Directory to output log files to}

\item{run_name}{Run name to save log files as}

\item{stdout_log}{(logical) Output logs to stdout()? Default true}

\item{thin}{Thinning rate for continuous parameter posterior samples}

\item{save_migration_history}{Logical value indicating whether tree samples are saved (i.e. whether .trees file is created)}

\item{migration_history_thin}{Thinning rate for tree samples to be saved}

\item{proposal_rates}{Relative rates of migration history and continuous parameter updates (migration history : coalescent rates : migration rates)}
}
\description{
Runs an MCMC chain using the MCMC method from Ewing et al. (2004) with Gamma-distributed priors
}
