% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitch.R
\name{fitch}
\alias{fitch}
\title{Fitch Algorithm}
\usage{
fitch(strphylo)
}
\arguments{
\item{strphylo}{Dated structured phylogenetic tree containing at least tip demes}
}
\description{
Computes the number of required migration events for a maximum parsimony migration history on a given structured phylogenetic tree
}
