% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_likelihoods.R
\name{local_DTA_likelihood}
\alias{local_DTA_likelihood}
\title{#' Local Structured Coalescent Likelihood
#'
#' Computes the structured coalescent likelihood locally between two specified events
#'
#'
#' @param ED Extended data representation of a phylogeny including migration history
#' @param coal_rate Vector of coalescent rates
#' @param bit_mig_mat Backward-in-time migration matrix for the phylogeny
#' @param event_ids Pair of event IDs to compute likelihood between
#' @param node_indices Node indices
#' @param deme_decomp Deme Decomp
#' @return Likelihood & log_likelihood between events
#'
#' @export}
\usage{
local_DTA_likelihood(st_labels, coal_rate, fit_mig_mat)
}
\arguments{
\item{st_labels}{subset of rows from an ED object giving all nodes in a subtree}

\item{coal_rate}{Vector of coalescent rates}

\item{fit_mig_mat}{Forwards-in-time migration matrix for the phylogeny}
}
\value{
Likelihood & log_likelihood between events
}
\description{

}
\details{

}
