% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_sampling.R
\name{local_DTA_subtree_proposal}
\alias{local_DTA_subtree_proposal}
\title{Local DTA}
\usage{
local_DTA_subtree_proposal(
  ED,
  st_labels,
  fit_rates,
  n_deme = nrow(fit_rates),
  node_indices = NodeIndicesC(ED),
  eigen_decomp = eigen(fit_rates),
  inverse_vecs = solve(eigen_decomp$vectors)
)
}
\arguments{
\item{ED}{ED representation of entire structured phylogenetic tree including virtual migrations needed to isolate subtree}

\item{st_labels}{ED representation of subtree only}

\item{fit_rates}{Forwards-in-time migration rates matrix}

\item{node_indices}{Vector of node indices. Element i gives the row of label i}

\item{eigen_decomp}{Eigendecomposition of fit_rates (output from eigen(fit_rates))}

\item{inverse_vecs}{Inverse of matrix of eigenvectors (output from solve(eigen(fit_rates)$vectors))}
}
\value{
List of proposed DTA migration histories with the same leaves as input ED
}
\description{
Updates the migration history on a subtree under the DTA model
}
