% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Local_likelihoods.R
\name{local_MTT_transition_kernel}
\alias{local_MTT_transition_kernel}
\title{Local MTT Likelihood}
\usage{
local_MTT_transition_kernel(
  ED,
  st_labels,
  bit_rates,
  node_indices = NodeIndicesC(ED),
  eigen_vals = eigen(bit_rates)$values,
  eigen_vecs = eigen(bit_rates)$vectors,
  inverse_vecs = solve(eigen_vecs)
)
}
\arguments{
\item{ED}{Structured phylogeny using internal data structure}

\item{st_labels}{subset of rows from an ED object giving all nodes in a subtree}

\item{bit_rates}{Backwards-in-time migration rates matrix}
}
\value{
Likelihood & log_likelihood between events
}
\description{
Computes the proposal kernel for a MultiTypeTree NodeRetype operator
}
