% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ewing_2004_proposals.R
\name{migration_pair_death}
\alias{migration_pair_death}
\title{Migration Pair Death MCMC Move}
\usage{
migration_pair_death(ED, n_deme, node_indices)
}
\arguments{
\item{ED}{Extended data object; matrix with columns Node ID, parent, child 1, child 2, deme, node age}

\item{n_deme}{Number of distinct demes in the population}
}
\value{
Updated extended data object with the proposal from the migration pair death move
}
\description{
Performs a Migration pair death move (Ewing et al. 2004). Deletes two
migration nodes if they lie between two edges in the same deme.
}
