% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vaughan_2015_MultiTypeTree_NodeRetype.R
\name{MTT_transition_kernel}
\alias{MTT_transition_kernel}
\title{MultiTypeTree Probability Density}
\usage{
MTT_transition_kernel(
  ED,
  bit_rates,
  node_indices = NodeIndicesC(ED),
  eigen_vals = eigen(bit_rates)$values,
  eigen_vecs = eigen(bit_rates)$vectors,
  inverse_vecs = solve(eigen_vecs)
)
}
\arguments{
\item{ED}{Extended data representation of phylogenetic tree and initial migration history}

\item{node_indices}{Vector of row indices corresponding to which row of ED corresponds to each node label}

\item{bit_mig_mat}{Matrix of initial backwards-in-time migration rates for the MCMC chain}

\item{st_labels}{Reduced extended data structure consisting of the rows of ED corresponding to coalescent nodes in the selected subtree only (local likelihood only)}
}
\value{
List consisting of the log.likelihood and likelihood of the structured genealogy
}
\description{
Computes the transition kernel for a MultiTypeTree NodeRetype operator for a full structured phylogeny
}
