% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstrphylo.R
\name{rstrphylo}
\alias{rstrphylo}
\title{Simulation of Heterochronous Structured Phylogenies}
\usage{
rstrphylo(
  n,
  d,
  tip_data = cbind(1:n, 0, rep_len(1:d, n)),
  coalescent_rates,
  migration_matrix
)
}
\arguments{
\item{n}{Number of samples}

\item{d}{Number of demes}

\item{tip_data}{nx3 matrix with first column giving tip labels, second column giving tip ages and third column giving tip demes}

\item{coalescent_rates}{Numeric vector giving coalescent rates (inverse effective population sizes) of each deme}

\item{migration_matrix}{Numeric matrix giving backwards-in-t migration rates between each pair of demes}
}
\value{
An object of class '\code{strphylo}'
}
\description{
Simulates a dated phylogenetic tree under the structured coalescent model
}
