% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA_sampling.R
\name{st_centre_dist}
\alias{st_centre_dist}
\title{Subtree Sampling}
\usage{
st_centre_dist(ED, st_radius, NI, st_child = NA, st_centre_loc = runif(1))
}
\arguments{
\item{ED}{ED representation of structured phylogenetic tree}

\item{st_radius}{Subtree radius}

\item{NI}{Vector of node indices. Element i gives the row of label i}

\item{st_child}{Child node of the subtree centre}

\item{st_centre_loc}{Value between 0 and 1 giving position between st_child and ED[st_child,2]}
}
\value{
List of proposed DTA migration histories with the same leaves as input ED
}
\description{
Samples a subtree up to patristic distance st_radius from a subtree centre
}
