% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihoods.R
\name{structured_coalescent_likelihood}
\alias{structured_coalescent_likelihood}
\title{Structured Coalescent likelihood}
\usage{
structured_coalescent_likelihood(strphylo, coal_rate, bit_mig_mat)
}
\arguments{
\item{strphylo}{Structured dated phylogeny}

\item{coal_rate}{Vector of coalescent rate in each deme (1 / effective population size)}

\item{bit_mig_mat}{Matrix of backwards-in-time migration rates between pairs of demes}
}
\value{
List containing log-likelihood and likelihood of the structured phylogenetic tree
}
\description{
Computes the likelihood of a structured phylogenetic tree under the structured coalescent model
}
